refactor: lambdaディレクトリ構造を呼び出し元ごとに整理（api_gateway/event_bridge/dynamodb_stream）およびテストも対応
