+++ lambda_functions/api_gateway/settings_api.py
