+++ lambda_functions/dynamodb_stream/notify_slack_stream.py
