+++ lambda_functions/event_bridge/tweet_monitor_batch.py
